+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added: reverse methods `findMatchesReverse`, `findLastMatch`, `lastIndexOf`, `findEndMatch`, `endsWith`
+
 ## `1.1.0` – 2026-Mar-22
 
 - Added: method `findStartMatch`